Added rounded buttons and centered text. - Text can now be aligned to a specific corner. - Added default colors to Colors class. - Added GUI Demo application. - Updated Clock and Calculator for new gui style.

+++ apps/compositor/contextmanager.cpp
@@ -14,8 +14,8 @@ ContextInfo* ContextManager::FindTargetContext(int x, int y)
     // Loop through all known contexts starting at the one on the front
     for(ContextInfo* c : contextList)
     {
-        if(x >= c->x && x <= c->x + c->width)       // Check if the coordinate x,y fits in the border
-            if(y >= c->y && y <= c->y + c->height)  // of the context. If so return as result.
+        if(x >= c->x && x <= c->x + (int)c->width)       // Check if the coordinate x,y fits in the border
+            if(y >= c->y && y <= c->y + (int)c->height)  // of the context. If so return as result.
                 return c;
     }
